+++ Changelog.md
@@ -2,7 +2,13 @@
 
 ## Changes
 
+### Auto Pass
+
+Clarified Auto Pass option description.
+
 ### Classic Era and TBC
+
+Added new option for "Classic Auto Pass" - a special version that only auto passes on items you *cannot* equip as opposed to the normal *cannot or should not* rule.
 
 Removed transmog options for these versions as transmog doesn't exist.